// src/generated/models/MTNRecipient.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mastercard Multi Token Network API
 * MTN (Multi Token Network) API is for customers and their consumers to - create account - buy/sell/hold cryptos - resolve alias and addresses - mint/transfer/burn tokenized deposits on blockchain
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: apisupport@mastercard.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MTNReceipt } from './MTNReceipt';
import {
    MTNReceiptFromJSON,
    MTNReceiptFromJSONTyped,
    MTNReceiptToJSON,
} from './MTNReceipt';

/**
 * Recipient for an Earmark
 * @export
 * @interface MTNRecipient
 */
export interface MTNRecipient {
    /**
     * Consumer account identifier.
     * @type {string}
     * @memberof MTNRecipient
     */
    accountAlias?: string;
    /**
     * The amount of asset.
     * @type {number}
     * @memberof MTNRecipient
     */
    amount?: number;
    /**
     * 
     * @type {Array<MTNReceipt>}
     * @memberof MTNRecipient
     */
    receipts?: Array<MTNReceipt>;
}

/**
 * Check if a given object implements the MTNRecipient interface.
 */
export function instanceOfMTNRecipient(value: object): boolean {
    return true;
}

export function MTNRecipientFromJSON(json: any): MTNRecipient {
    return MTNRecipientFromJSONTyped(json, false);
}

export function MTNRecipientFromJSONTyped(json: any, ignoreDiscriminator: boolean): MTNRecipient {
    if (json == null) {
        return json;
    }
    return {
        
        'accountAlias': json['accountAlias'] == null ? undefined : json['accountAlias'],
        'amount': json['amount'] == null ? undefined : json['amount'],
        'receipts': json['receipts'] == null ? undefined : ((json['receipts'] as Array<any>).map(MTNReceiptFromJSON)),
    };
}

export function MTNRecipientToJSON(value?: MTNRecipient | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'accountAlias': value['accountAlias'],
        'amount': value['amount'],
        'receipts': value['receipts'] == null ? undefined : ((value['receipts'] as Array<any>).map(MTNReceiptToJSON)),
    };
}