// src/generated/models/MTNTransferOperation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mastercard Multi Token Network API
 * MTN (Multi Token Network) API is for customers and their consumers to - create account - buy/sell/hold cryptos - resolve alias and addresses - mint/transfer/burn tokenized deposits on blockchain
 *
 * The version of the OpenAPI document: 0.0.2
 * Contact: apisupport@mastercard.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MTNBlockchainTransaction } from './MTNBlockchainTransaction';
import {
    MTNBlockchainTransactionFromJSON,
    MTNBlockchainTransactionFromJSONTyped,
    MTNBlockchainTransactionToJSON,
} from './MTNBlockchainTransaction';
import type { MTNTokenIdentifier } from './MTNTokenIdentifier';
import {
    MTNTokenIdentifierFromJSON,
    MTNTokenIdentifierFromJSONTyped,
    MTNTokenIdentifierToJSON,
} from './MTNTokenIdentifier';
import type { MTNTokenOperation } from './MTNTokenOperation';
import {
    MTNTokenOperationFromJSON,
    MTNTokenOperationFromJSONTyped,
    MTNTokenOperationToJSON,
} from './MTNTokenOperation';

/**
 * A request body example to transfer bank tokens operation.
 * @export
 * @interface MTNTransferOperation
 */
export interface MTNTransferOperation extends MTNTokenOperation {
    /**
     * Amount of tokens to transfer.
     * @type {number}
     * @memberof MTNTransferOperation
     */
    amount: number;
    /**
     * Consumer account identifier.
     * @type {string}
     * @memberof MTNTransferOperation
     */
    to: string;
    /**
     * Consumer account identifier.
     * @type {string}
     * @memberof MTNTransferOperation
     */
    from: string;
}

/**
 * Check if a given object implements the MTNTransferOperation interface.
 */
export function instanceOfMTNTransferOperation(value: object): boolean {
    if (!('amount' in value)) return false;
    if (!('to' in value)) return false;
    if (!('from' in value)) return false;
    return true;
}

export function MTNTransferOperationFromJSON(json: any): MTNTransferOperation {
    return MTNTransferOperationFromJSONTyped(json, false);
}

export function MTNTransferOperationFromJSONTyped(json: any, ignoreDiscriminator: boolean): MTNTransferOperation {
    if (json == null) {
        return json;
    }
    return {
        ...MTNTokenOperationFromJSONTyped(json, ignoreDiscriminator),
        'amount': json['amount'],
        'to': json['to'],
        'from': json['from'],
    };
}

export function MTNTransferOperationToJSON(value?: Omit<MTNTransferOperation, 'operationId'|'status'|'blockchainTransactions'> | null): any {
    if (value == null) {
        return value;
    }
    return {
        ...MTNTokenOperationToJSON(value),
        'amount': value['amount'],
        'to': value['to'],
        'from': value['from'],
    };
}